fix compilation on FreeBSD

+++ sys/dev/netmap/netmap_mem2.c
@@ -1492,8 +1492,8 @@ netmap_mem2_rings_create(struct netmap_adapter *na)
 				netmap_mem_set_ring(na->nm_mem, ring->slot, ndesc, 0);
 			}
 		        /* ring info */
-		        *(uint16_t *)&ring->ringid = kring->ring_id;
-		        *(uint16_t *)&ring->dir = 0;
+		        *(uint16_t *)(uintptr_t)&ring->ringid = kring->ring_id;
+		        *(uint16_t *)(uintptr_t)&ring->dir = 0;
 		}
 	}
